fixed to_f

+++ test/item_test.rb
@@ -50,7 +50,7 @@ class ItemTest < Minitest::Test
   end
 
   def test_it_can_set_unit_price_to_dollars
-    assert_instance_of Float, @item.unit_price
+    assert_instance_of BigDecimal, @item.unit_price
     assert_equal 10.99, @item.unit_price
   end
 end
